🚚 zb: Move UDS socket type impls to their own module
